+++ pkg/apic/apiserver/models/management/v1alpha1/model_access_request_spec_quota.go
@@ -12,6 +12,7 @@ package v1alpha1
 // AccessRequestSpecQuota Quota information for accessing the api. (management.v1alpha1.AccessRequest)
 type AccessRequestSpecQuota struct {
 	// The limit of the allowed quota for the access request.
-	Limit    float32 `json:"limit"`
+	// GENERATE: The following code has been modified after code generation
+	Limit    float64 `json:"limit"`
 	Interval string  `json:"interval"`
 }
